refactor(ui): move queryParam to util
