+++ ui/src/app/shared/services/cron-workflow-service.ts
@@ -1,4 +1,5 @@
 import {CronWorkflow, CronWorkflowList} from '../../../models';
+import {Utils} from '../utils';
 import requests from './requests';
 
 export class CronWorkflowService {
@@ -11,7 +12,7 @@ export class CronWorkflowService {
 
     public list(namespace: string, labels: string[] = []) {
         return requests
-            .get(`api/v1/cron-workflows/${namespace}?${this.queryParams({labels}).join('&')}`)
+            .get(`api/v1/cron-workflows/${namespace}?${Utils.queryParams({labels}).join('&')}`)
             .then(res => res.body as CronWorkflowList)
             .then(list => list.items || []);
     }
@@ -37,26 +38,5 @@ export class CronWorkflowService {
 
     public resume(name: string, namespace: string) {
         return requests.put(`api/v1/cron-workflows/${namespace}/${name}/resume`).then(res => res.body as CronWorkflow);
-    }
-
-    private queryParams(filter: {labels?: Array<string>}) {
-        const queryParams: string[] = [];
-        const labelSelector = this.labelSelectorParams(filter.labels);
-        if (labelSelector.length > 0) {
-            queryParams.push(`listOptions.labelSelector=${labelSelector}`);
-        }
-
-        return queryParams;
-    }
-
-    private labelSelectorParams(labels?: Array<string>) {
-        let labelSelector = '';
-        if (labels && labels.length > 0) {
-            if (labelSelector.length > 0) {
-                labelSelector += ',';
-            }
-            labelSelector += labels.join(',');
-        }
-        return labelSelector;
     }
 }
