+++ ui/src/app/shared/services/workflow-template-service.ts
@@ -1,4 +1,5 @@
 import * as models from '../../../models';
+import {Utils} from '../utils';
 import requests from './requests';
 
 export class WorkflowTemplateService {
@@ -11,7 +12,7 @@ export class WorkflowTemplateService {
 
     public list(namespace: string, labels: string[]) {
         return requests
-            .get(`api/v1/workflow-templates/${namespace}?${this.queryParams({labels}).join('&')}`)
+            .get(`api/v1/workflow-templates/${namespace}?${Utils.queryParams({labels}).join('&')}`)
             .then(res => res.body as models.WorkflowTemplateList)
             .then(list => list.items || []);
     }
@@ -29,26 +30,5 @@ export class WorkflowTemplateService {
 
     public delete(name: string, namespace: string) {
         return requests.delete(`api/v1/workflow-templates/${namespace}/${name}`);
-    }
-
-    private queryParams(filter: {labels?: Array<string>}) {
-        const queryParams: string[] = [];
-        const labelSelector = this.labelSelectorParams(filter.labels);
-        if (labelSelector.length > 0) {
-            queryParams.push(`listOptions.labelSelector=${labelSelector}`);
-        }
-
-        return queryParams;
-    }
-
-    private labelSelectorParams(labels?: Array<string>) {
-        let labelSelector = '';
-        if (labels && labels.length > 0) {
-            if (labelSelector.length > 0) {
-                labelSelector += ',';
-            }
-            labelSelector += labels.join(',');
-        }
-        return labelSelector;
     }
 }
