+++ ui/src/app/shared/utils.ts
@@ -1,6 +1,7 @@
 import {Observable} from 'rxjs';
 import * as models from '../../models';
 import {NODE_PHASE} from '../../models';
+import {Pagination} from './pagination';
 
 const managedNamespaceKey = 'managedNamespace';
 const currentNamespaceKey = 'current_namespace';
@@ -118,5 +119,76 @@ export const Utils = {
     // return a namespace, never return null/undefined, defaults to "default"
     getNamespaceWithDefault(namespace: string) {
         return this.managedNamespace || namespace || this.currentNamespace || 'default';
+    },
+
+    queryParams(filter: {
+        namespace?: string;
+        name?: string;
+        namePrefix?: string;
+        phases?: Array<string>;
+        labels?: Array<string>;
+        minStartedAt?: Date;
+        maxStartedAt?: Date;
+        pagination?: Pagination;
+        resourceVersion?: string;
+    }) {
+        const queryParams: string[] = [];
+        const fieldSelector = this.fieldSelectorParams(filter.namespace, filter.name, filter.minStartedAt, filter.maxStartedAt);
+        if (fieldSelector.length > 0) {
+            queryParams.push(`listOptions.fieldSelector=${fieldSelector}`);
+        }
+        const labelSelector = this.labelSelectorParams(filter.phases, filter.labels);
+        if (labelSelector.length > 0) {
+            queryParams.push(`listOptions.labelSelector=${labelSelector}`);
+        }
+        if (filter.pagination) {
+            if (filter.pagination.offset) {
+                queryParams.push(`listOptions.continue=${filter.pagination.offset}`);
+            }
+            if (filter.pagination.limit) {
+                queryParams.push(`listOptions.limit=${filter.pagination.limit}`);
+            }
+        }
+        if (filter.namePrefix) {
+            queryParams.push(`namePrefix=${filter.namePrefix}`);
+        }
+        if (filter.resourceVersion) {
+            queryParams.push(`listOptions.resourceVersion=${filter.resourceVersion}`);
+        }
+        return queryParams;
+    },
+
+    fieldSelectorParams(namespace?: string, name?: string, minStartedAt?: Date, maxStartedAt?: Date) {
+        let fieldSelector = '';
+        if (namespace) {
+            fieldSelector += 'metadata.namespace=' + namespace + ',';
+        }
+        if (name) {
+            fieldSelector += 'metadata.name=' + name + ',';
+        }
+        if (minStartedAt) {
+            fieldSelector += 'spec.startedAt>' + minStartedAt.toISOString() + ',';
+        }
+        if (maxStartedAt) {
+            fieldSelector += 'spec.startedAt<' + maxStartedAt.toISOString() + ',';
+        }
+        if (fieldSelector.endsWith(',')) {
+            fieldSelector = fieldSelector.substr(0, fieldSelector.length - 1);
+        }
+        return fieldSelector;
+    },
+
+    labelSelectorParams(phases?: Array<string>, labels?: Array<string>) {
+        let labelSelector = '';
+        if (phases && phases.length > 0) {
+            labelSelector = `workflows.argoproj.io/phase in (${phases.join(',')})`;
+        }
+        if (labels && labels.length > 0) {
+            if (labelSelector.length > 0) {
+                labelSelector += ',';
+            }
+            labelSelector += labels.join(',');
+        }
+        return labelSelector;
     }
 };
